Add some developer docs for working with Loris (#918)
* Only fetch Loris config from AWS if not already present

This is for running in dev mode, if you're testing local changes
to config -- by copying the file into the container, you can skip
trying to fetch it from AWS.

* Add some developer docs for Loris

+++ loris/run_loris.sh
@@ -4,9 +4,16 @@ set -o errexit
 set -o nounset
 set -o xtrace
 
-echo "Fetching config from AWS..."
-aws s3 ls s3://$INFRA_BUCKET/$CONFIG_KEY
-aws s3 cp s3://$INFRA_BUCKET/$CONFIG_KEY /opt/loris/etc/loris2.conf
+CONF_FILE=/opt/loris/etc/loris.conf
+
+if [[ -f "$CONF_FILE" ]]
+then
+  echo "Config is already present; skipping fetch from AWS..."
+else
+  echo "Fetching config from AWS..."
+  aws s3 ls s3://$INFRA_BUCKET/$CONFIG_KEY
+  aws s3 cp s3://$INFRA_BUCKET/$CONFIG_KEY /opt/loris/etc/loris2.conf
+fi
 
 echo "=== config ==="
 cat /opt/loris/etc/loris2.conf
